requirements.txt
fastapi
uvicorn[standard]
prometheus-client 
python-multipart
pydantic-settings
gTTs
edge-tts